Fix RBAC permissions for the Antctl ClusterRole (#5166)
* Fix RBAC permissions for the Antctl ClusterRole

The ClusterRole definition was no longer up-to-date, with some incorrect
permissions and some missing permissions. As a consequence, it could not
be used to run some antctl commands, such as supportbundle.

We fix the permissions and modify the Antctl e2e tests so that they use
a Kubeconfig file generated for the antctl ServiceAccount, instead of
the admin Kubeconfig file. Hopefully, this will help keep the
ClusterRole definition up-to-date in the future.

A few other improvements were implemented in that process:

* antctl now uses the typed SystemBundle K8s clientset, instead of a raw
  REST client. This helps define the correct RBAC permissions for
  antctl. Previous permissions were indeed incorrect as they used the
  "post" verb (instead of the "create" verb), which is only correct for
  non-resource endpoints.
* supportbundle unit tests now use an in-memory filesystem, to avoid
  writing test outputs to the local machine.
* the antctl ClusterRole is giving access to a few extra endpoints
  ("/metrics", "/debug/pprof/*") to increse the usefulness of the antctl
  proxy command.

Fixes #5136

* Antctl test e2e improvements

* Don't use ~ (home dir) in antctl e2e tests with Kind

'~' is not expanded with docker exec

Signed-off-by: Antonin Bas <abas@vmware.com>