[VM Agent] Bugfix: antrea-agent failed to delete ExternalNode (#5191)
The issue is seen on a RHEL 8.4 VM on azure cloud, which is configured with
dhclient to manage the network interface. The root cause is antrea-agent fails
to recover the IP/Routes from host internal interface to uplink after
ExternalNode is deleted, because the added IP/Routes is deleted or conflicted
with dhclient configuration. Then in the continuous retry with ExternalNode
deletion, antrea-agent is blocking at the precheck on the existentance of host
internal interface always returns true as the uplink's name is already
recovered.

Signed-off-by: wenyingd <wenyingd@vmware.com>