Fix TestController_RotateCertificates unit test
The test is flaky because the computed rotation deadline is sometimes
<7s, due to the NotAfter certificate field being truncated at the
second. We can resolve the issue by also truncating the fake clock
initial time.

We also use the controller's clock time when validating
certificates. Because some tests use a fake clock, the time that was
used was incorrect, leading to some error logs - and potentially test
failures.

Fixes #5234

Signed-off-by: Antonin Bas <abas@vmware.com>