Upgrade OVS to 2.17.7 (#5231)
This is the latest patch version available for OVS 2.17.

We also fix a bug in the Dockerfile which installs OVS using .deb
packages built from source. The command that was used would try to
upgrade the openvswitch-* packages if a more recent version was
available in upstream repos, which would cause the installation to
fail. For now, we prefer to use the packages that we build from source,
as we have more granular control over which version gets installed. This
issue is fixed by using apt instead of dpkg to install the .deb packages
(this is supported in "recent" apt versions).

Thanks @Dyanngg for reporting the issue.

Signed-off-by: Antonin Bas <abas@vmware.com>