[e2e] Support skipping deploying Antrea
When developing something, it should be common that Antrea is already
deployed and we just want to validate it with e2e tests. For example, I
usually deploy Antrea using helm which is easier to set the
configurations. But running e2e always deployed Antrea and required the
manifest yaml to be present in the home directory of control plane node,
which adds some burden to running e2e tests.

This commit adds a flag "--deploy-antrea" to support skipping deploying
Antrea. It defaults to true to keep the same behavior as before. When we
just want to run e2e tests and not to redeploy Antrea. We can set
"--deploy-antrea=false".

Signed-off-by: Quan Tian <qtian@vmware.com>