Use new virtual time for TestController_RotateCertificates
Use a virtual time in the future. This is because
certutil.NewSelfSignedCACert uses real wall-clock time to generate the
root certificate, and there is a possibility that the NotBefore field
for the CA certificate is after the "now" virtual time. This can in turn
cause certificate validation to fail and the unit test to fail.

Signed-off-by: Antonin Bas <abas@vmware.com>