Handle unexpected error when the test is aborted
When the conformance test is aborted by Jenkins, the results log
is actually an incomplete file, we need also to check if there is a
`complete` text for e2e test result. Otherwise, the result will show
`All tests passed.` when the test is actually running and aborted
unexpectedly, which is misleading.

Signed-off-by: Lan Luo <luola@vmware.com>