Stop pulling Antrea images in ci/kind/test-e2e-kind.sh
The images should not be pulled, as we want to use the ones we have
built locally. The reason why this has not been an issue so far for CI
is that all jobs us the "coverage" images, which are not available in
the Docker registry (but the logs show error messages because of the
failed docker pull commands).

Signed-off-by: Antonin Bas <abas@vmware.com>