Add "X-Load-Balancing-Endpoint-Weight" header to AntreaProxy Service healthcheck
This commit adds the new header "X-Load-Balancing-Endpoint-Weight" to AntreaProxy.
The header will be returned from the service health and its value will represent
the number of local Endpoints. This header enables weighted load balancing and
offers a faster alternative to parsing JSON from the content body.

The addition of this header is necessary for achieving parity with kube-proxy.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>