Fix GKE conformance job (#5310)
* Stop exporting KUBECONFIG in ci/test-conformance-gke.sh

This is causing the cloud-antrea-gke-cleanup Jenkins job to fail (even
though the cluster is deleted properly), because for some reason "gcloud
container clusters delete" tries to create the Kubeconfig file (and the
directory does not exist):

ERROR: (gcloud.container.clusters.delete) Unable to create private file [/home/ubuntu/jenkins/out/gke/kubeconfig]: [Errno 1] Operation not permitted: '/home/ubuntu/jenkins/out/gke/kubeconfig'

This issue was introduced in #4862. It was not caught during testing
because the cleanup job, which is triggered by the main conformance job,
was not triggered using the correct parameters. Current build parameters
(ANTREA_REPO and ANTREA_GIT_REVISION) need to be propagated to the
cleanup job to enable proper testing before merges. We update the cloud
job definitions so that current-parameters is set to true when
triggering the cleanup jobs.

Signed-off-by: Antonin Bas <abas@vmware.com>

* Fix GKE CI job

The gke-gcloud-auth-plugin binary is now required (in addition to the
main gcloud binary).

As a result, we make some changes to the test-conformance-gke.sh script.
The script now takes a '--gcloud-sdk-path' flag, which points to the
installation directory of the gcloud SDK. The bin directory from the SDK
is then added to the path.

When running the script locally, one can omit the '--gcloud-sdk-path'
flag, as long as the PATH already include the SDK bin directory (which
is the default hen following standard installation instructions).

Signed-off-by: Antonin Bas <abas@vmware.com>

* Remove 'source ~/.bashcrc' command from Jenkins job definitions

On Ubuntu, this command will not have any effect as the .bashrc file
exits immediately for non-interactive shells.

Signed-off-by: Antonin Bas <abas@vmware.com>

* Update cleanup jobs

Signed-off-by: Antonin Bas <abas@vmware.com>

* Fix GKE conformance job

The job was only deleting Pods in the kube-system Namespace after
installing Antrea. However, GKE has been introducing new Namespaces
(e.g., gmp-system), which are also running non-hostNetwork Pods. If we
do not delete ALL non-hostNetwork Pods in ALL Namespaces after
installing Antrea, it can lead to IP address reuse, and a broken GKE
cluster.

Signed-off-by: Antonin Bas <abas@vmware.com>

* Skip legacy NetworkPolicy tests in GKE conformance job

And switch back to the new "Netpol" test suite.
The legacy jobs can still pass, but take a very long time because of a
change in K8s v1.27:
https://github.com/kubernetes/kubernetes/commit/d8428c6fb1c09d7113ae0413b08c04a83dce5285

The legacy NetworkPolicy tests are also removed as of K8s v1.28.

Signed-off-by: Antonin Bas <abas@vmware.com>

---------

Signed-off-by: Antonin Bas <abas@vmware.com>