Stop using /bin/sh for OVS commands
We are currently using the shell (/bin/sh) to run ovs-appctl and
ovs-ofctl commands, but there does not seem to be a reason to do so.

Instead, we can invoke the binary directly with exec.Command and pass
the arguments as a slice.

Before this change, ovsVSwitchdUDS would return
"/var/run/openvswitch/ovs-vswitchd.*.ctl" as a fallback value in case
the PID of ovs-vswitchd could not be determined. The shell would then
expand the wildcard patterns. I now believe that it is better to return
an error in that case. We want to make sure that we use a valid file
always (and returning the glob pattern would cause an error if multiple
files match). If we run into issues because of this approach, we can
update the code to expand the glob pattern programmatically and, if a
single path matches, we can return that path.

Signed-off-by: Antonin Bas <abas@vmware.com>