Support Pod secondary interfaces on VLAN network (#5365)
This commit adds support for connecting Pod secondary interfaces to a
VLAN network. A Pod secondary interface configured with the VLAN
networkType will be connected to the secondary network OVS bridge, and
the specified VLAN tag will be set on the OVS port.
This commit mostly follows the existing SR-IOV secondary network
implementation, in which a Pod controller configures Pod secondary
interfaces based on the local Pod events. It thus inherits the existing
limitations, like secondary interface configuration is not persisted on
the Node and may get lost after antrea-agent restarts.

Signed-off-by: Jianjun Shen <shenj@vmware.com>