Support creating extra Docker networks in kind tests (#5416)
Add an --extra-networks option to kind-setup.sh to create extra Docker
bridge networks which will be connected by each kind worker node. The
--extra-networks option cannot be specified together with the --subnets
option.
Add an --extra-network option to test-e2e-kind.sh which will create a
single extra network for the kind test cluster. The option does not work
with the hybrid encap mode (which runs kind-setup.sh with the  --subnets
option). The extra network will be required for the e2e tests of
secondary VLAN network, which create a secondary OVS bridge on each
worker node, and need an extra host interface connected on the bridge.

Signed-off-by: Jianjun Shen <shenj@vmware.com>