Do not attempt to join Windows agents to memberlist cluster
Windows Nodes should not join the memberlist cluster as all features
relying on it is not only supported on Windows. Currently an agent only
runs the memberlist module and listen to the memberlist port when
required so attempts to join Windows agents are destined to fail,
leading to misleading error logs.

This commit fixes it by ignoring Windows Nodes when joining agents.

Signed-off-by: Quan Tian <qtian@vmware.com>