Handle stale LabelIdentities created for empty Pod labels correctly after upgrade
This is a follow-up on PR #5404 which intends to address issue #5403.
In upgrade cases, Antrea controller will still receive LabelIdentity add events
for LabelIdentities created by previous controller, which it needs to handle
correctly. Also, the stale controller logic needs to be updated to ensure that
LabelIdentities created with old normalization method are cleaned up after upgrade.

Signed-off-by: Dyanngg <dingyang@vmware.com>