Fix burst of PacketInQueue (#5456)
The burst was not effective because the number of tokens was hardcoded
to 1, leading to 10ms delay when processing every successive DNS
response. For example, in some scenario it takes a client 10 DNS queries
to get a valid DNS response, the later 9 DNS responses would be delayed
90ms in total. This affected the DNS resolution delay when a Pod has any
FQDN policy applied.

This commit fixes the burst setting of the PacketInQueue. The connection
time of accessing a FQDN is reduced from 134ms to 41ms according to my
test.

Signed-off-by: Quan Tian <qtian@vmware.com>