Fix flaky TestCluster_Run unit test
The test was flaky on machine with a limited amount of CPU, and I was
able to reproduce locally by setting GOMAXPROCS to 1.

If the Node object is created after starting the informers and waiting
for them to be synced, then the Node may not be "visible" to the lister
when the ExternalIPPool is synced. So we need to create the Node when we
instantiate the fake K8s client, before starting the informers.

Signed-off-by: Antonin Bas <abas@vmware.com>