Fix flaky TestFlowAggregator_Run unit test (#5515)
The test was flaky because sometimes not all expectations were met. For
example:

> mockPodStore.EXPECT().Run(gomock.Any())

The reason is that the FlowAggregator Run method can return "too
quickly", while some of its goroutines are still in the process of being
started. Rather than make these expectations optional, which I think is
a bit hacky, we leverage WaitGroups in Run to ensure that all child
goroutines complete before returning from Run.

Signed-off-by: Antonin Bas <abas@vmware.com>