[Test] Make TestBatchCreatePods more robust
This e2e validates that when creating a batch of Pods, no file
descriptor is leaked by the antrea-agent process. In order to do that,
it compares the sets of FDs before creating the Pods and after creating
the Pods. However, in some rare cases, we observe that the test fails
because of some temporary FDs that have been allocated, e.g., because of
incoming CNI requests. We make the test more robust by asserting that
the new set of FDs (after batch creation) eventually becomes a subset of
the old set of FDs (before batch creation). The polling logic performs
the check every 100ms and only needs a small timeout (2s), as we never
expect the check to take more than a couple of iterations. In most
cases, it will succeed on the first iteration.

Fixes #5567

Signed-off-by: Antonin Bas <abas@vmware.com>