Add checksum of windows config files as Deployment annotations (#5545)
When the ConfigMap is generated with suffix by kustomize, it will
generate a manifest with a new ConfigMap like 'antrea-windows-config-***'.
When a new manifest is applied, it will leave stale ConfigMaps.
To avoid stale ConfigMaps, here we add checksum of windows configuration
files as annotations for the antrea-agent Deployment. If the ConfigMap's
contents are updated, the annotations will change and the Deployment will
be automatically updated by K8s. This is similar to what we have in Linux
manifests.

Signed-off-by: Kumar Atish <atish.iaf@gmail.com>