Add a script for drafting release
Creating releases is an error-prone operation. Instead of relying on
people to be meticulous, we can automate the process.

The commit adds a script to draft a release in Github for a given
version, based on the changelog. It still leaves the final click to
people before it's proven reliable.

To make Github include a "Contributors" section with an avatar list of
all the mentioned contributors, the author links are removed.

Signed-off-by: Quan Tian <qtian@vmware.com>