Use netip.Addr for FlowExporter (#5532)
* Improvements to FlowExporter benchmarks

Also add BenchmarkConnStore, meant to measure memory usage of the
connection store (and all the connections).

* Replace net.IP with netip.Addr in FlowExporter implementation

The net/netip package was introduced in Go 1.18. Compared to the
existing net.IP type, the netip.Addr type takes less memory (especially
for IPv6 addresses), is immutable, and is comparable so it supports ==
and can be used as a map key.

In our case, this means that the flowexporter.Connection type takes a
little less memory, and that the flowexporter.Tuple can be used directly
as the flowexporter.ConnectionKey (no need to generate a string for use
as the map key). Overall, this leads to higher performance (decreased
CPU and memory usage) for the FlowExporter.

There is potential for further improvement, as several key functions
take as parameters both the flowexporter.Connection and
flowexporter.ConnectionKey objects, which is no longer really required.

See #5271

* Bump up github.com/ti-mo/conntrack to v0.5

This release uses netip.Addr instead of net.IP to represent connections.

Signed-off-by: Antonin Bas <abas@vmware.com>