Fix antrea-ubi image build
Starting with Go 1.21, the runtime system needs to have a version of
glibc "compatible" with the version available at build time. The ubi8
image comes with a very old version of glibc compared to the golang
builder, and the Antrea binaries no longer run.

To address this, we use the ubi8 image to build Antrea, which means we
need to install Go manually.

Fixes #5722

Signed-off-by: Antonin Bas <abas@vmware.com>