Automatically upgrade Go version in Jenkins CI (#5388) (#5448)
Everytime the Jenkins jobs are triggered, the jobs will check the golang
version on the testbed and compare with golang version in go.mod. If the
golang version is different, the jobs will automatically download and
install newer golang. If a new patch version is available, it will also
install it.

Signed-off-by: hjiajing <hjiajing@vmware.com>