Update Windows documentation
The user-facing Windows document is updated as follows:
* containerd will be the only supported container runtime on Windows
  Nodes, Docker support (that includes cri-dockerd, Docker Desktop) is
  officially deprecated.
* we split the "Installation as Pod" section into 2 new sections (one
  for containerd, one for Docker).
* installation methods are re-ordered, starting with the recommended /
  most common one: Installation as a Pod for containerd, Installation as
  Windows services, Installation as a Pod for Docker (deprecated).

The main idea is to make the installation instructions clearer and
easier to follow for a "normal" user, deploying the latest Antrea
version on a recent K8s cluster. In this situation, the recommended
installation method is to use antrea-windows-containerd.yml or
antrea-windows-containerd-with-ovs.yml, which leverage HostProcess
Windows Pods.

There is still a lot of room for future improvements. I suggest that for
Antrea v2.0 we drop all references to Docker and kube-proxy from this
document.

The "Antrea Management on Windows" section is also dropped from the
Windows design document, as it was stale information.

For #5624 #5630

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>