Delete OVS port and flows before releasing Pod IP
Pod IP is recorded as OVS port attribute and used in Pod specific flows.
If releasing Pod IP before deleting OVS port and flows, it could happen
that multiple Pod ports and flows reference to the same Pod IP as the
same time in some corner cases and lead to corrupted network.

This commit ensures all resources that reference to Pod IP are deleted
before releasing the IP.

Signed-off-by: Quan Tian <qtian@vmware.com>