Upgrade to a more recent version of markdownlint-cli
From 0.31.1 to 0.38.0

We have not upgraded the version used in CI for a long time, and the new
version includes some additional checks such as MD051 link-fragments
(Link fragments should be valid), which helps avoid deadlinks in the
documentation (internal links in a given document).

Contributors can install the new version of markdownlint-cli using npm
or Brew (macOS).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>