Add script for migrating from other CNIs to Antrea (#5677)
Add a YAML file antrea-migrator.yml to help migrate clusters
with other CNIs to Antrea. It will restart all Pods in-place.

A new DaemonSet "antrea-migrator" is responsible for cleaning
up configuration from the previous CNI (e.g., iptables rules), as
well as "restarting" all Pods on all Nodes, to ensure that they are
connected to the Antrea network. To restart Pods with minimum
downtime and without causing them to be rescheduled, it uses
the `crictl stopp` command to restart the Pods in-place.

Fixes #5578 

Signed-off-by: hjiajing <hjiajing@vmware.com>