Don't create tunnel interface (antrea-tun0) when using Wireguard
When using Wireguard for inter-Node traffic, there is no need to create
the OVS tunnel port (antrea-tun0) interface, as traffic will go through
antrea-wg0, which is managed by Linux directly.

The exception is when using Multi-cluster (in gateway mode), as
antrea-tun0 is then still used for cross-cluster traffic, for example to
tunnel traffic from the source Node to the local gateway Node.

Fixes #5869

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>