Add instructions to install Antrea in encap mode in AKS
AKS has supported a "Bring Your Own CNI" mode for a while now:
https://learn.microsoft.com/en-us/azure/aks/use-byo-cni?tabs=azure-cli

With this support, Antrea can easily be installed in encap mode in AKS
clusters, with no limitation (unlike for EKS where there is no more
connectivity from the control-plane to the Nodes). The only requirement
is to enable Antrea NodeIPAM.

This support relies on the
[Konnectivity Service](https://kubernetes.io/docs/tasks/extend-kubernetes/setup-konnectivity/),
which is deployed by default in "recent" AKS versions, and guarantees
connectivity from the control-plane to the Nodes, regardless of which
CNI is used.

Fixes #5817

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>