Remove invalid flow when WireGuard is enabled
After merging #5885, the OVS tunnel port is no longer created when
enabling WireGuard, as it is not being used. However, we were still
trying to install flows referencing the tunnel port. Even though the
port was non-existent, flow creation would succeed, but the condition
matching on the tunnel port was being dropped silently. This would lead
to invalid datapath behavior. We update the code to prevent installing
these invalid flows when WireGuard is enabled.

Fixes #5905

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>