Optimize Docker builds
We use Docker cache mounts to accelerate Docker builds with Go.
With cache mounts, we can cache package dependencies and the Go build
cache in a persistent fashion (i.e., persists across builds).

As a minor optimization, we also use a bind mount for go.mod & go.sum,
to avoid a COPY statement and an extra layer. We could do the same
thing for the antrea source directory, but it would add extra complexity
to the Dockerfiles, with no clear benefit (~2s per build).

With these changes, an incremental builds using
./hack/build-antrea-linux-all.sh takes less than a minute (instead of
more than 5 minutes).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>